feat(#3280):  Add support for pipeline templates (#3281)
* feat(#3266): Add API to programmatically create pipelines

* Fix checkstyle

* feat(#3280): Add support for pipeline templates

* Fix validation

* Remove obsolete stylesheet

* Fix dependency

* Add default template to setup

* feat(#3292): Show adapter and pipeline configuration in UI (#3293)

* feat(#3292): Add code tab to adapter view

* Show pipeline code in UI

* Add header

* Fix checkstyle

* Fix checkstyle

* Fix template visitor

* Remove empty css class

+++ streampipes-wrapper-standalone/src/main/java/org/apache/streampipes/wrapper/standalone/StreamPipesNotificationSink.java
@@ -84,7 +84,7 @@ public abstract class StreamPipesNotificationSink extends StreamPipesDataSink {
     // convert input given in minutes to seconds
     // this is later used to determine if a notification should be sent
     this.silentPeriodInSeconds = parameters.extractor()
-                                           .singleValueParameter(KEY_SILENT_PERIOD, Long.class) * 60
+                                           .singleValueParameter(KEY_SILENT_PERIOD, Integer.class) * 60
     ;
   }
